Создал новые компоненты «Список отзывов» и «Отзыв». С помощью компонента «Список отзывов», обновил компонент страницы «Room»
